fixed #53

+++ app/Models/Banking/Account.php
@@ -71,14 +71,14 @@ class Account extends Model
     }
 
     /**
-     * Convert opening balance to float.
+     * Convert opening balance to double.
      *
      * @param  string  $value
      * @return void
      */
     public function setOpeningBalanceAttribute($value)
     {
-        $this->attributes['opening_balance'] = (float) $value;
+        $this->attributes['opening_balance'] = (double) $value;
     }
 
     /**
