+++ database/migrations/2017_09_01_000000_create_taxes_table.php
@@ -16,7 +16,7 @@ class CreateTaxesTable extends Migration
             $table->increments('id');
             $table->integer('company_id');
             $table->string('name');
-            $table->integer('rate');
+            $table->double('rate', 15, 4);
             $table->boolean('enabled');
             $table->timestamps();
             $table->softDeletes();
