+++ database/migrations/2017_10_11_000000_create_bill_totals_table.php
@@ -24,7 +24,7 @@ class CreateBillTotalsTable extends Migration
             $table->integer('bill_id');
             $table->string('code')->nullable();
             $table->string('name');
-            $table->float('amount', 15, 4);
+            $table->double('amount', 15, 4);
             $table->integer('sort_order');
             $table->timestamps();
             $table->softDeletes();
